Issue #11532 - Refactor test trait
